Adding original samples from Starbucks
These will need to be filtered and edited for presentation worthiness --
there's a lot of noise. This has authentication and association captures
for attwifi, cablewifi, and twcwifi (three commerical open wifi points
available at this particular Starbucks), as well as a sample association
to my own Android AP and a couple failed attempts at other random APs.

The differences in the DHCP sequences and the capture portals between
the three public APs are kind of neat, as well as the levels of ARP
traffic between them.